Created Cache, Group, Hom, StrTrie, and WordTrie
Group is the principle part of this program as it stores the semi-decision process, the algorithm for the word problem as well as many other methods for groups. The rest of these are either used for data storage or are retired parts of the program that are no longer used.

Hom stands for homomorphism and was used early on in understanding word and letter objects. Cache was an attempt to store all of the information by brute force when a different semi-decision process was being employed. StrTrie was the base logic for WordTrie which stores word objects in particular in a Trie objects so as to make them easier to search.